Move all special forms to result macros
This commit consists mostly of cutting special-form methods of the compiler and pasting them over as macros. Likewise, a lot of private compiler methods have become helper functions. The following changes are made:

- Replace `@special` with `@result_macro`
- Replace `self` with `compiler`, and change some method calls to function calls accordingly
- Remove some leading underscores in names
- Adjust spacing and the order of definitions
- Add Outshine headers for Emacs in comments
- Add a module docstring
